Move components into component folder and fix some descrpitions
